+++ src/workers/BlockMintingWorker.ts
@@ -8,6 +8,7 @@ import BasicWorker from './BasicWorker';
 import Settings from '../types/Settings';
 import CryptoCompareWSInitializer from '../services/CryptoCompareWSInitializer';
 import PolygonIOPriceInitializer from '../services/PolygonIOPriceInitializer';
+import KaikoPriceStreamInitializer from '../services/KaikoPriceStreamInitializer';
 
 @injectable()
 class BlockMintingWorker extends BasicWorker {
@@ -16,6 +17,7 @@ class BlockMintingWorker extends BasicWorker {
   @inject(BlockMinter) blockMinter!: BlockMinter;
   @inject(CryptoCompareWSInitializer) cryptoCompareWSInitializer!: CryptoCompareWSInitializer;
   @inject(PolygonIOPriceInitializer) polygonIOPriceInitializer!: PolygonIOPriceInitializer;
+  @inject(KaikoPriceStreamInitializer) kaikoPriceStreamInitializer!: KaikoPriceStreamInitializer;
 
   enqueue = async <T>(params: T, opts?: Bull.JobsOptions): Promise<Bull.Job<T> | undefined> => {
     const isLocked = await this.connection.get(this.settings.jobs.blockCreation.lock.name);
@@ -63,6 +65,12 @@ class BlockMintingWorker extends BasicWorker {
     });
 
     this.cryptoCompareWSInitializer.apply().catch((err: Error) => {
+      newrelic.noticeError(err);
+      this.logger.error(err);
+      process.exit(1);
+    });
+
+    this.kaikoPriceStreamInitializer.apply().catch((err: Error) => {
       newrelic.noticeError(err);
       this.logger.error(err);
       process.exit(1);