+++ test/services/FeedProcessor.test.ts
@@ -36,6 +36,7 @@ describe('FeedProcessor', () => {
     CoinmarketcapHistoDayFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.CoinmarketcapHistoDayFetcher>,
     BEACPIAverageFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.BEACPIAverageFetcher>,
     OnChainDataFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.OnChainDataFetcher>,
+    KaikoPriceStreamFetcher: null as unknown as sinon.SinonStubbedInstance<fetchers.KaikoPriceStreamFetcher>,
   };
 
   let feedProcessor: FeedProcessor;
@@ -55,6 +56,7 @@ describe('FeedProcessor', () => {
     mockedFetchers.CoinmarketcapHistoDayFetcher = sinon.createStubInstance(fetchers.CoinmarketcapHistoDayFetcher);
     mockedFetchers.BEACPIAverageFetcher = sinon.createStubInstance(fetchers.BEACPIAverageFetcher);
     mockedFetchers.OnChainDataFetcher = sinon.createStubInstance(fetchers.OnChainDataFetcher);
+    mockedFetchers.KaikoPriceStreamFetcher = sinon.createStubInstance(fetchers.KaikoPriceStreamFetcher);
 
     const container = new Container();
 
@@ -98,6 +100,7 @@ describe('FeedProcessor', () => {
     container.bind(fetchers.CoingeckoPriceFetcher).toConstantValue(mockedFetchers.CoingeckoPriceFetcher as any);
     container.bind(fetchers.BEACPIAverageFetcher).toConstantValue(mockedFetchers.BEACPIAverageFetcher as any);
     container.bind(fetchers.OnChainDataFetcher).toConstantValue(mockedFetchers.OnChainDataFetcher as any);
+    container.bind(fetchers.KaikoPriceStreamFetcher).toConstantValue(mockedFetchers.KaikoPriceStreamFetcher as any);
 
     container.bind(FeedProcessor).toSelf();
 
@@ -343,5 +346,34 @@ describe('FeedProcessor', () => {
     } catch (err) {
       expect(err).to.be.false;
     }
+  });
+
+  it('returns leafs for feeds with KaikoPriceStreamFetcher fetcher', async () => {
+    const feeds: Feeds = {
+      'BTC-USD': {
+        discrepancy: 1,
+        precision: 2,
+        inputs: [
+          {
+            fetcher: {
+              name: 'KaikoPriceStream',
+              params: {
+                fsym: 'BTC',
+                tsym: 'USD',
+              } as any,
+            },
+          },
+        ],
+      },
+    };
+
+    mockedFetchers.KaikoPriceStreamFetcher.apply.resolves(38123);
+
+    const leaves = await feedProcessor.apply(10, feeds);
+
+    expect(leaves[0]).to.be.an('array').with.lengthOf(1);
+
+    expect(leaves[0][0].label).to.equal('BTC-USD');
+    expect(LeafValueCoder.decode(leaves[0][0].valueBytes, leaves[0][0].label)).is.a('number').that.equal(38123);
   });
 });