+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- add `KaikoSpotPrice` http fetcher
+- add `KaikoPriceStream` stream fetcher
 
 ## [5.1.1] - 2021-09-03
 ### Fixed